pyproject.toml
[project]
name = "cofe-shop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.56",
    "celery[redis]>=5.5.3",
    "django>=5.2.7",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.9.0",
    "django-csp>=4.0",
    "django-extensions>=4.1",
    "django-filter>=25.2",
    "django-health-check>=3.20.0",
    "django-import-export>=4.3.12",
    "django-mptt>=0.18.0",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "django-robots>=6.1",
    "django-silk>=5.4.3",
    "django-storages>=1.14.6",
    "django-treebeard>=4.7.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "markdown>=3.9",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.11",
    "pytest-django>=4.11.1",
    "python-decouple>=3.8",
    "sentry-sdk>=2.42.1",
    "uvicorn>=0.38.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-debug-toolbar>=6.0.0",
    "faker>=37.11.0",
    "flake8>=7.3.0",
]
